standalone_systests: isolate dirent tests at dirent.c
Move the dirent-specific test cases from semihost.c to dirent.c. This
way we can test semihost OPs on systems that do not support dirent yet,
such as qemu on Windows.
